requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
Flask==3.1.0
Flask-Bcrypt==1.0.1
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.91.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.5
openai==0.28.1
packaging==25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
